+++ cli/cli.pro
@@ -8,13 +8,14 @@ QT -= gui
 SOURCES += src/main.cpp \
     src/commandargsparser.cpp \
     src/clisubtitlesdownloader.cpp \
-    src/console.cpp
+    src/console.cpp \
+    src/climain.cpp
     
 HEADERS += src/qnapiclicommand.h \
     src/commandargsparser.h \
     src/clisubtitlesdownloader.h \
-    src/namespace_tr.h \
-    src/console.h
+    src/console.h \
+    src/climain.h
 
 RESOURCES += res/resources.qrc
 
@@ -42,11 +43,8 @@ macx {
 
 win32 {
     CONFIG += nostrip
-
     RC_FILE = ../win32/qnapi.rc
-
     TARGET = qnapic
-
     target.path = ../win32/out
     INSTALLS += target
 }
