+++ cli/src/main.cpp
@@ -12,170 +12,14 @@
 **
 *****************************************************************************/
 
-#include "clisubtitlesdownloader.h"
+#include "climain.h"
 #include "commandargsparser.h"
 #include "console.h"
 #include "libqnapi.h"
-#include "namespace_tr.h"
 #include "qnapiclicommand.h"
-#include "subtitlelanguage.h"
 
 #include <QCoreApplication>
-#include <QDir>
 #include <QTranslator>
-#include <QVariant>
-
-#include <signal.h>
-#include <iostream>
-
-namespace Main {
-
-Q_DECLARE_NAMESPACE_TR(Main)
-
-void installTranslation(QCoreApplication &app, QTranslator *translator,
-                        const QNapiConfig &config) {
-  QString uiLanguage = LibQNapi::uiLanguage(config.generalConfig());
-  translator->load("qnapi_" + uiLanguage, ":/translations");
-  app.installTranslator(translator);
-}
-
-void printHeader(const Console &c) {
-  c.printLine(tr("QNapi %1 (Qt version %2), %3")
-                  .arg(LibQNapi::displayableVersion())
-                  .arg(qVersion())
-                  .arg(LibQNapi::webpageUrl()));
-  c.printLine();
-}
-
-void printHelp(const Console &c) {
-  QString formats =
-      LibQNapi::subtitleFormatsRegistry()->listFormatNames().join(",");
-  QString binaryFileName =
-      QFileInfo(LibQNapi::appExecutableFilePath).fileName();
-
-  c.printLine(
-      tr("QNapi is distributed under the GNU General Public License v2."));
-  c.printLine();
-  c.printLine(tr("Syntax: %1 [options] [list of files]").arg(binaryFileName));
-  c.printLine(tr("Available options:"));
-  c.printLine(
-      tr("   -q, --quiet                Download subtitles quietly without "
-         "showing"));
-  c.printLine(
-      tr("                              any messages or windows (implies -d)"));
-  c.printLine();
-  c.printLine(
-      tr("   -s, --show-list            Show a list of subtitles (works only "
-         "with -c)"));
-  c.printLine(
-      tr("   -d, --dont-show-list       Do not show a list of subtitles (works "
-         "only with -c)"));
-  c.printLine();
-  c.printLine(tr("   -l, --lang [code]          Preferred subtitles language"));
-  c.printLine(
-      tr("   -lb,--lang-backup [code]   Alternative subtitles language"));
-  c.printLine(tr("   -f, --format [format]      Select target subtitles file "
-                 "format (%1)")
-                  .arg(formats));
-  c.printLine(tr(
-      "   -e, --extension [ext]      Select target subtitles file extension"));
-  c.printLine();
-  c.printLine(tr("   -h, --help                 Show help text"));
-  c.printLine(tr(
-      "   -hl,--help-languages       List of available subtitles languages"));
-  c.printLine();
-}
-
-void printHelpLanguages(const Console &c, const QNapiConfig &config) {
-  c.printLine(
-      tr("List of languages recognized by QNapi, including corresponding"));
-  c.printLine(tr("two-letter language codes:"));
-  c.printLine();
-
-  SubtitleLanguage L, LB;
-  QStringList langs = L.listLanguages();
-
-  foreach (QString lang, langs) {
-    L.setLanguage(lang);
-    c.printLine(QString(" %1 - %2").arg(L.toTwoLetter()).arg(lang));
-  }
-
-  L.setLanguage(config.generalConfig().language());
-  LB.setLanguage(config.generalConfig().backupLanguage());
-
-  c.printLine();
-  c.printLine(tr("Current default subtitles language: %1 (%2)")
-                  .arg(L.toFullName())
-                  .arg(L.toTwoLetter()));
-
-  if (LB.toFullName().isEmpty()) {
-    c.printLine(tr("No alternative subtitles language has been set"));
-  } else {
-    c.printLine(tr("Current alternative subtitles language: %1 (%2)")
-                    .arg(LB.toFullName())
-                    .arg(LB.toTwoLetter()));
-  }
-}
-
-int processCommand(QVariant cliCommand, const QNapiConfig &config) {
-  const Console c(config.generalConfig().quietBatch());
-
-  printHeader(c);
-
-  using namespace QNapiCliCommand;
-  if (cliCommand.canConvert<DownloadSubtitles>()) {
-    QStringList movieFilePaths =
-        cliCommand.value<DownloadSubtitles>().movieFilePaths;
-    return CliSubtitlesDownloader::downloadSubtitlesFor(c, movieFilePaths,
-                                                        config);
-  } else if (cliCommand.canConvert<ShowHelpLanguages>()) {
-    printHelpLanguages(Console(), config);
-    return 0;
-  } else {
-    printHelp(Console());
-    return 0;
-  }
-}
-
-void sigHandler(int sig) {
-  Q_UNUSED(sig);
-
-  std::cout << std::endl
-            << tr("QNapi: deleting temporary files...").toStdString()
-            << std::endl;
-
-  const QNapiConfig config = LibQNapi::loadConfig();
-  QString tmpPath = config.generalConfig().tmpPath();
-  QDir tmpDir(tmpPath);
-
-  QStringList filters;
-  filters << "QNapi-*-rc";
-  filters << "QNapi.*.tmp";
-
-  QFileInfoList files = tmpDir.entryInfoList(filters);
-
-  foreach (QFileInfo file, files) { QFile::remove(file.filePath()); }
-
-  std::cout << tr("QNapi: finished.").toStdString() << std::endl;
-
-  exit(666);
-}
-
-void regSignal() {
-#ifdef Q_OS_WIN
-  signal(SIGTERM, sigHandler);
-  signal(SIGINT, sigHandler);
-#else
-  struct sigaction sa;
-  memset(&sa, 0, sizeof(struct sigaction));
-  sigemptyset(&sa.sa_mask);
-  sa.sa_handler = sigHandler;
-  sigaction(SIGTERM, &sa, 0);
-  sigaction(SIGINT, &sa, 0);
-#endif
-}
-
-}  // namespace Main
 
 int main(int argc, char **argv) {
   LibQNapi::init(argv[0]);
@@ -184,10 +28,10 @@ int main(int argc, char **argv) {
 
   QCoreApplication cliApp(argc, argv);
 
-  Main::regSignal();
+  CliMain::regSignal();
 
   QTranslator qnapiTranslator;
-  Main::installTranslation(cliApp, &qnapiTranslator, config);
+  CliMain::installTranslation(cliApp, &qnapiTranslator, config);
 
   QStringList tailArgs =
       cliApp.arguments().mid(1, cliApp.arguments().size() - 1);
@@ -196,13 +40,12 @@ int main(int argc, char **argv) {
 
   if (parseResult.is<CommandArgsParser::ParsedCommand>()) {
     auto parsedCommand = parseResult.as<CommandArgsParser::ParsedCommand>();
-    return Main::processCommand(parsedCommand.command,
-                                parsedCommand.refinedConfig);
-
+    return CliMain::processCommand(parsedCommand.command,
+                                   parsedCommand.refinedConfig);
   } else {
     const Console c;
-    Main::printHeader(c);
-    c.printLine(Main::tr("Command line argument parsing error:"));
+    CliMain::printHeader(c);
+    c.printLine(CliMain::tr("Command line argument parsing error:"));
     c.printLine(parseResult.as<QString>());
     return -1;
   }
