+++ gui/gui.pro
@@ -21,7 +21,8 @@ SOURCES += src/main.cpp \
     src/qcumber/qsingleapplication.cpp \
     src/qnapiopendialog.cpp \
     src/qnapiapp.cpp \
-    src/qnapicli.cpp 
+    src/qnapicli.cpp \ 
+    src/guimain.cpp
 
 HEADERS += src/forms/frmprogress.h \
     src/forms/frmlistsubtitles.h \
@@ -43,7 +44,8 @@ HEADERS += src/forms/frmprogress.h \
     src/qnapiopendialog.h \
     src/qnapiapp.h \
     src/qnapicli.h \
-    src/qnapicommand.h
+    src/qnapicommand.h \
+    src/guimain.h
 
 FORMS += ui/frmprogress.ui \
     ui/frmlistsubtitles.ui \
@@ -77,11 +79,10 @@ macx {
     SOURCES += src/utils/infoplistdockicon.cpp
     HEADERS += src/utils/infoplistdockicon.h
 
-    LIBS += -framework CoreFoundation
-
     TARGET = QNapi
     DESTDIR = ../macx/
 
+    LIBS += -framework CoreFoundation
     QMAKE_MACOSX_DEPLOYMENT_TARGET = 10.7
     QMAKE_CXXFLAGS_X86_64 = -mmacosx-version-min=10.7
     ICON = ../macx/qnapi.icns
@@ -95,14 +96,10 @@ macx {
 
 win32 {
     CONFIG += nostrip
-
     SOURCES += src/qcumber/qinterprocesschannel_win32.cpp
     HEADERS += src/qcumber/qinterprocesschannel_win32.h
-
     RC_FILE = ../win32/qnapi.rc
-
     TARGET = qnapi
-
     target.path = ../win32/out
     INSTALLS += target
 }
