+++ libqnapi/src/namespace_tr.h
